moved interface utils out of its own folder fixed typo and minor formatting
